+++ components/menu/src/ItemGroup.tsx
@@ -1,0 +1,29 @@
+import { getPropsSlot } from '../../_util/props-util';
+import { computed, defineComponent } from 'vue';
+import PropTypes from '../../_util/vue-types';
+import { useInjectMenu } from './hooks/useMenuContext';
+
+export default defineComponent({
+  name: 'AMenuItemGroup',
+  props: {
+    title: PropTypes.VNodeChild,
+  },
+  slots: ['title'],
+  setup(props, { slots }) {
+    const { prefixCls } = useInjectMenu();
+    const groupPrefixCls = computed(() => `${prefixCls.value}-item-group`);
+    return () => {
+      return (
+        <li onClick={e => e.stopPropagation()} class={groupPrefixCls.value}>
+          <div
+            title={typeof props.title === 'string' ? props.title : undefined}
+            class={`${groupPrefixCls.value}-title`}
+          >
+            {getPropsSlot(slots, props, 'title')}
+          </div>
+          <ul class={`${groupPrefixCls.value}-list`}>{slots.default?.()}</ul>
+        </li>
+      );
+    };
+  },
+});
