+++ components/menu/style/dark.less
@@ -1,7 +1,8 @@
 .@{menu-prefix-cls} {
   // dark theme
-  &-dark,
-  &-dark &-sub {
+  &&-dark,
+  &-dark &-sub,
+  &&-dark &-sub {
     color: @menu-dark-color;
     background: @menu-dark-bg;
     .@{menu-prefix-cls}-submenu-title .@{menu-prefix-cls}-submenu-arrow {
@@ -19,8 +20,7 @@
   }
 
   &-dark &-inline&-sub {
-    background: @menu-dark-submenu-bg;
-    box-shadow: 0 2px 8px fade(@black, 45%) inset;
+    background: @menu-dark-inline-submenu-bg;
   }
 
   &-dark&-horizontal {
@@ -31,8 +31,13 @@
   &-dark&-horizontal > &-submenu {
     top: 0;
     margin-top: 0;
+    padding: @menu-item-padding;
     border-color: @menu-dark-bg;
     border-bottom: 0;
+  }
+
+  &-dark&-horizontal > &-item:hover {
+    background-color: @menu-dark-item-active-bg;
   }
 
   &-dark&-horizontal > &-item > a::before {
@@ -41,7 +46,8 @@
 
   &-dark &-item,
   &-dark &-item-group-title,
-  &-dark &-item > a {
+  &-dark &-item > a,
+  &-dark &-item > span > a {
     color: @menu-dark-color;
   }
 
@@ -77,7 +83,8 @@
   &-dark &-submenu-title:hover {
     color: @menu-dark-highlight-color;
     background-color: transparent;
-    > a {
+    > a,
+    > span > a {
       color: @menu-dark-highlight-color;
     }
     > .@{menu-prefix-cls}-submenu-title,
@@ -95,6 +102,10 @@
     background-color: @menu-dark-item-hover-bg;
   }
 
+  &-dark&-dark:not(&-horizontal) &-item-selected {
+    background-color: @menu-dark-item-active-bg;
+  }
+
   &-dark &-item-selected {
     color: @menu-dark-highlight-color;
     border-right: 0;
@@ -102,14 +113,19 @@
       border-right: 0;
     }
     > a,
-    > a:hover {
+    > span > a,
+    > a:hover,
+    > span > a:hover {
       color: @menu-dark-highlight-color;
     }
+
+    .@{menu-prefix-cls}-item-icon,
     .@{iconfont-css-prefix} {
       color: @menu-dark-selected-item-icon-color;
-    }
-    .@{iconfont-css-prefix} + span {
-      color: @menu-dark-selected-item-text-color;
+
+      + span {
+        color: @menu-dark-selected-item-text-color;
+      }
     }
   }
 
@@ -122,7 +138,8 @@
   &-dark &-item-disabled,
   &-dark &-submenu-disabled {
     &,
-    > a {
+    > a,
+    > span > a {
       color: @disabled-color-dark !important;
       opacity: 0.8;
     }
