--- components/new-menu/src/ItemGroup.tsx
@@ -1,16 +1,0 @@
-import { getPropsSlot } from '../../_util/props-util';
-import { defineComponent } from 'vue';
-
-export default defineComponent({
-  name: 'AMenuItemGroup',
-  setup(props, { slots }) {
-    return () => {
-      return (
-        <li>
-          {getPropsSlot(slots, props, 'title')}
-          <ul>{slots.default?.()}</ul>
-        </li>
-      );
-    };
-  },
-});
