--- components/new-menu/src/Menu.tsx
@@ -1,10 +1,0 @@
-import { defineComponent } from 'vue';
-
-export default defineComponent({
-  name: 'AMenu',
-  setup(props, { slots }) {
-    return () => {
-      return <div>{slots.default?.()}</div>;
-    };
-  },
-});
