--- components/new-menu/src/MenuItem.tsx
@@ -1,10 +1,0 @@
-import { defineComponent } from 'vue';
-
-export default defineComponent({
-  name: 'AMenuItem',
-  setup(props, { slots }) {
-    return () => {
-      return <li>{slots.default?.()}</li>;
-    };
-  },
-});
