+++ components/old-menu/style/dark.less
@@ -1,8 +1,7 @@
 .@{menu-prefix-cls} {
   // dark theme
-  &&-dark,
-  &-dark &-sub,
-  &&-dark &-sub {
+  &-dark,
+  &-dark &-sub {
     color: @menu-dark-color;
     background: @menu-dark-bg;
     .@{menu-prefix-cls}-submenu-title .@{menu-prefix-cls}-submenu-arrow {
@@ -20,7 +19,8 @@
   }
 
   &-dark &-inline&-sub {
-    background: @menu-dark-inline-submenu-bg;
+    background: @menu-dark-submenu-bg;
+    box-shadow: 0 2px 8px fade(@black, 45%) inset;
   }
 
   &-dark&-horizontal {
@@ -31,13 +31,8 @@
   &-dark&-horizontal > &-submenu {
     top: 0;
     margin-top: 0;
-    padding: @menu-item-padding;
     border-color: @menu-dark-bg;
     border-bottom: 0;
-  }
-
-  &-dark&-horizontal > &-item:hover {
-    background-color: @menu-dark-item-active-bg;
   }
 
   &-dark&-horizontal > &-item > a::before {
@@ -46,8 +41,7 @@
 
   &-dark &-item,
   &-dark &-item-group-title,
-  &-dark &-item > a,
-  &-dark &-item > span > a {
+  &-dark &-item > a {
     color: @menu-dark-color;
   }
 
@@ -83,8 +77,7 @@
   &-dark &-submenu-title:hover {
     color: @menu-dark-highlight-color;
     background-color: transparent;
-    > a,
-    > span > a {
+    > a {
       color: @menu-dark-highlight-color;
     }
     > .@{menu-prefix-cls}-submenu-title,
@@ -102,10 +95,6 @@
     background-color: @menu-dark-item-hover-bg;
   }
 
-  &-dark&-dark:not(&-horizontal) &-item-selected {
-    background-color: @menu-dark-item-active-bg;
-  }
-
   &-dark &-item-selected {
     color: @menu-dark-highlight-color;
     border-right: 0;
@@ -113,19 +102,14 @@
       border-right: 0;
     }
     > a,
-    > span > a,
-    > a:hover,
-    > span > a:hover {
+    > a:hover {
       color: @menu-dark-highlight-color;
     }
-
-    .@{menu-prefix-cls}-item-icon,
     .@{iconfont-css-prefix} {
       color: @menu-dark-selected-item-icon-color;
-
-      + span {
-        color: @menu-dark-selected-item-text-color;
-      }
+    }
+    .@{iconfont-css-prefix} + span {
+      color: @menu-dark-selected-item-text-color;
     }
   }
 
@@ -138,8 +122,7 @@
   &-dark &-item-disabled,
   &-dark &-submenu-disabled {
     &,
-    > a,
-    > span > a {
+    > a {
       color: @disabled-color-dark !important;
       opacity: 0.8;
     }
